[09/01] 더미 데이터 생성 기능 추가
아이템 더미 데이터 생성 기능 추가

+++ banto/src/test/java/com/example/banto/FavoriteDAOTest.java
@@ -34,7 +34,7 @@ class FavoriteDAOTest {
 	@Transactional
 	void successTest() {
 		System.out.println("성공 메소드 - input : 1");
-		int id = 1;
+		long id = 1;
 		// Data
 		
 		// When
